Put listeners on drag handle to prevent event interference
Moving the drag and drop listeners will prevent them from interfering
with other listeners. In particular, putting the drag and drop listeners
on the li element will cause a full page reload when clicking on the
edit icon link inside the li element because the edit icon's Link
component won't be given the chance to handle the click event.